Move soft lock tests to dedicated file
